Moved model-related stuff to backend.ml module
